requirements.txt
torch>=2.0
numpy>=1.24
pygame-ce>=2.5
tqdm>=4.65
